HiGlassViewConfig Item-Type; Refactoring HiGlassContainer
