+++ src/encoded/static/components/item-pages/components/HiGlass/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+export { HiGlassContainer, HiGlassConfigurator } from './HiGlassContainer';
+export { HiGlassPlainContainer } from './HiGlassPlainContainer';
+export { HiGlassTabView } from './HiGlassTabView';
+export { HiGlassLocalStorage } from './HiGlassLocalStorage';
