+++ src/encoded/static/components/item-pages/components/index.js
@@ -22,7 +22,7 @@ export { TabbedView } from './TabbedView';
 export { ItemDetailList, Detail } from './ItemDetailList';
 export { AuditTabView } from './AuditTabView';
 export { AttributionTabView } from './AttributionTabView';
-export { HiGlassTabView, HiGlassContainer, HiGlassConfigurator } from './HiGlassContainer';
+export { HiGlassTabView, HiGlassContainer, HiGlassPlainContainer, HiGlassConfigurator } from './HiGlass';
 export { WorkflowDetailPane } from './WorkflowDetailPane';
 export { WorkflowNodeElement } from './WorkflowNodeElement';
 export { FlexibleDescriptionBox, FlexibleCharacterCountBox } from './FlexibleDescriptionBox';
